requirements.txt
# requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-dotenv==1.0.1
aiohttp==3.9.3
jinja2==3.1.3
python-multipart==0.0.9
pandas==2.2.1
